% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rdatamaps.R
\name{rdatamaps}
\alias{rdatamaps}
\title{<Add Title>}
\usage{
rdatamaps(data = NULL, scope = NULL, fills = NULL, projection = NULL,
  dataUrl = NULL, keyCol = NULL, clng = NULL, clat = NULL,
  cwidth = NULL, cheight = NULL, popup = NULL, popupVar = NULL,
  width = NULL, height = NULL)
}
\description{
<Add Description>
}

