% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rdatamaps.R
\name{renderRdatamaps}
\alias{renderRdatamaps}
\title{Widget render function for use in Shiny}
\usage{
renderRdatamaps(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

